+++ src/Http/Controllers/UserGuideController.php
@@ -10,7 +10,9 @@ use App\Http\Controllers\Controller;
 class UserGuideController extends Controller
 {
     protected $layout = 'layouts.app';
+
     public function index() {
+        $this->authorize('viewAny', UserGuide::class);
         $userGuides = UserGuide::with('module')->paginate(10);
         return view('moduleuserguide::userguides.index', [
             'userGuides' => $userGuides,
@@ -19,6 +21,7 @@ class UserGuideController extends Controller
     }
 
     public function create() {
+        $this->authorize('create', UserGuide::class);
         $modules = Module::all();
         return view('moduleuserguide::userguides.create', [
             'modules' => $modules,
@@ -27,6 +30,7 @@ class UserGuideController extends Controller
     }
 
     public function store(UserGuideRequest $request) {
+        $this->authorize('create', UserGuide::class);
         $data = $request->validated();
 
         // handle file uploads
@@ -44,6 +48,7 @@ class UserGuideController extends Controller
     }
 
     public function edit(UserGuide $userGuide) {
+        $this->authorize('update', $userGuide);
         $modules = Module::all();
         return view('moduleuserguide::userguides.edit', [
             'userGuide' => $userGuide,
@@ -53,6 +58,7 @@ class UserGuideController extends Controller
     }
 
     public function update(UserGuideRequest $request, UserGuide $userGuide) {
+        $this->authorize('update', $userGuide);
         $data = $request->validated();
 
         $files = $userGuide->files ?? [];
@@ -69,7 +75,8 @@ class UserGuideController extends Controller
     }
 
     public function destroy(UserGuide $userGuide) {
-        // delete files
+        $this->authorize('delete', $userGuide);
+
         if($userGuide->files){
             foreach($userGuide->files as $file){
                 Storage::delete($file);
@@ -78,5 +85,4 @@ class UserGuideController extends Controller
         $userGuide->delete();
         return redirect()->route('user-guides.index')->with('success','User Guide deleted successfully!');
     }
-
 }
